Generated Array of 10 Random Numbers for Tenzies in App Component
This commit adds the functionality to generate an array of 10 random numbers between 1 and 6 to the App component for the Tenzies game. The numbers are generated using the Math.random function and can be used to determine the players' moves in the game. The array is stored in a global variable and can be accessed by all functions in the game. This commit also includes tests to ensure that the array is generated correctly and contains the correct values.